+++ CHANGES.rst
@@ -5,6 +5,7 @@ Tunfish pocpoc changelog
 development
 ===========
 - Add virtual circuitry based on VXLAN through the `tunfish-join.sh` program
+- Add IP connectivity probing after network setup. Improve console output.
 
 2018-03-30 0.1.0
 ================