+++ CHANGES.rst
@@ -7,6 +7,7 @@ development
 - Add `tunfish-join.sh` program to establish virtual layer 2 circuitry through VXLAN
 - Add IP connectivity probing after network setup. Improve console output.
 - Rename "sandbox" directory to "environment"
+- Naming things. Polishing. Refactoring. Documentation.
 
 2018-03-30 0.1.0
 ================