+++ CHANGES.rst
@@ -1,5 +1,5 @@
 #########################
-Tunfish sandbox changelog
+Tunfish Sandbox Changelog
 #########################
 
 development